netmount client: Check sequence number of response
Each packet (request and response) contains a sequence number.
If the client does not receive a response, it repeats the original
request with the same sequence number. The server thus recognizes that
it is the same request, repeats the response without re-executing
the request (re-executing some operations may cause a problem).

The response contains the same sequence number as the request. This is
how the client knows that it is a response to the last request and not
a delayed response to a previous repeated request.

The sequence number check was missing in the client. This commit adds
it.